# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Source Accounting API",
    "description": "Free general chat and paid annual Agent workflows. All research outputs are drafts.",
    "version": "0.7.0"
  },
  "paths": {
...
    "/api/v1/memos": {
      "get": {
        "tags": [
          "memos"
        ],
        "summary": "List Memos",
        "operationId": "list_memos_api_v1_memos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "memos"
        ],
        "summary": "Create Manual",
        "operationId": "create_manual_api_v1_memos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memos/{memo_id}": {
      "get": {
        "tags": [
          "memos"
        ],
        "summary": "Memo Detail",
        "operationId": "memo_detail_api_v1_memos__memo_id__get",
        "parameters": [
          {
            "name": "memo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "memos"
        ],
        "summary": "Edit",
        "operationId": "edit_api_v1_memos__memo_id__put",
        "parameters": [
          {
            "name": "memo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memos/{memo_id}/revisions": {
      "get": {
        "tags": [
          "memos"
        ],
        "summary": "Revisions",
        "operationId": "revisions_api_v1_memos__memo_id__revisions_get",
        "parameters": [
          {
            "name": "memo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memos/{memo_id}/reviews": {
      "post": {
        "tags": [
          "memos"
        ],
        "summary": "Review",
        "operationId": "review_api_v1_memos__memo_id__reviews_post",
        "parameters": [
          {
            "name": "memo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memos/{memo_id}/export": {
      "get": {
        "tags": [
          "memos"
        ],
        "summary": "Export",
        "operationId": "export_api_v1_memos__memo_id__export_get",
        "parameters": [
          {
            "name": "memo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memo Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "md",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoCreate": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "maxLength": 200000,
            "title": "Body",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workspace_id",
          "title"
        ],
        "title": "MemoCreate"
      },
      "MemoUpdate": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "maxLength": 200000,
            "title": "Body"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "title",
          "body"
        ],
        "title": "MemoUpdate"
      },
...
      "ReviewCreate": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "note": {
            "type": "string",
            "maxLength": 3000,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision"
        ],
        "title": "ReviewCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}